New detection code

+++ tests/test_wrongkeys.sh
@@ -16,17 +16,24 @@ cat $TEST/test_input.txt | $PROG --format=json "stdin:rtlwmbus" \
       ApWater apator162   88888888 00000000000000000000000000000001 \
       Vatten  multical21  76348799 28F64A24988064A079AA2C807D6102AF \
       Wasser  supercom587 77777777 5065747220486F6C79737A6577736B6A \
-      > $TEST/test_output.txt
+      > $TEST/test_output.txt 2> $TEST/test_stderr.txt
 
-cat $TEST/test_output.txt | grep -v '{"media' > $TEST/test_response.txt
+if [ -s $TEST/test_output.txt ]
+then
+    echo "Bad no stdout expected! But got bytes anyway!"
+    echo "ERROR: $TESTNAME"
+    TESTRESULT="ERROR"
+    exit 1
+fi
 
 cat <<EOF > $TEST/test_expected.txt
+Started config rtlwmbus on stdin listening on any
 (wmbus) decrypted content failed check, did you use the correct decryption key? Ignoring telegram.
 (wmbus) decrypted payload crc failed check, did you use the correct decryption key? Ignoring telegram.
 (wmbus) decrypted content failed check, did you use the correct decryption key? Ignoring telegram.
 EOF
 
-diff $TEST/test_expected.txt $TEST/test_response.txt
+diff $TEST/test_expected.txt $TEST/test_stderr.txt
 if [ "$?" = "0" ]
 then
     echo "OK: $TESTNAME"
